Upgrade base pom 15.3 -> 15.7
This change was necessary to get the required jersey server version to
work with the most recent AWS SDK. However, it required a handful of
internal changes to get it to behave properly. Most importantly, these were
- Major version bump from jersey 1 to jersey 2, which included migrating
  between groupId's. It only required one piece of code-rewriting, which
  was reasonably small.
- Migrating groupId's from com.codahale.metrics to io.dropwizard.metrics.
  There was no real code changes involved, but Maven couldn't detect
  artifact conflicts between these two because they had different
  groupId's even though they were the same artifact. This meant that
  the dependency resolution had to be done by hand, so I just
  migrated to the newer version.
