+++ BaragonData/src/main/java/com/hubspot/baragon/auth/BaragonAuthBundle.java
@@ -1,5 +1,7 @@
 package com.hubspot.baragon.auth;
 
+import org.glassfish.jersey.server.ResourceConfig;
+
 import io.dropwizard.Bundle;
 import io.dropwizard.setup.Bootstrap;
 import io.dropwizard.setup.Environment;
@@ -12,6 +14,7 @@ public class BaragonAuthBundle implements Bundle {
 
   @Override
   public void run(Environment environment) {
-    environment.jersey().getResourceConfig().getResourceFilterFactories().add(BaragonAuthResourceFilterFactory.class);
+    ResourceConfig resourceConfig = environment.jersey().getResourceConfig();
+    resourceConfig.register(BaragonAuthFeature.class);
   }
 }
